Make EPS pip installable (from source directory)
Reorganise everything to use distutils.  This allows installation
and running tests using pip (or setup.py).  A basic thumb rule that
was followed was to use `electrumpersonalserver` in code, and use
`electrum-personal-server` where a user might need to type in
terminal (e.g. pip package/script/directory name).

ChangeLog:
- package directory reorganisation
  - renamed electrumpersonalserver → server
  - move all sub-packages under electrumpersonalserver/
  - change a few imports to explicitly import the required names
- remove all scripts (`server.py`, `rescan-script.py`, `*.bat`), as
  cross-platform scripts are now supported via the `console_script`
  entry point feature provided by `distutils`, all corresponding code
  has been moved to `electrumpersonalserver.server.common`
- `pytest` is now integrated in `setup.py`
- MANIFEST.in: include docs, certs, and sample config in source
  distribution
- include sample config in pip installation

TODO:
- proper argument parsing in scripts
- correctly use the installed certs in scripts
- log properly
