+++ electrumpersonalserver/__init__.py
@@ -1,10 +1,2 @@
-
-
-from electrumpersonalserver.merkleproof import convert_core_to_electrum_merkle_proof
-from electrumpersonalserver.jsonrpc import JsonRpc, JsonRpcError
-from electrumpersonalserver.hashes import (to_bytes, sha256, bh2u,
-          script_to_scripthash, get_status_electrum, bfh, hash_encode,
-          hash_decode, Hash, hash_merkle_root, hash_160, script_to_address,
-          address_to_script, address_to_scripthash, bytes_fmt)
-from electrumpersonalserver.transactionmonitor import TransactionMonitor, import_addresses, ADDRESSES_LABEL
-from electrumpersonalserver.deterministicwallet import parse_electrum_master_public_key, DeterministicWallet
+import electrumpersonalserver.bitcoin
+import electrumpersonalserver.server
