+++ electrumpersonalserver/bitcoin/secp256k1_deterministic.py
@@ -1,4 +1,4 @@
-from bitcoin.secp256k1_main import *
+from electrumpersonalserver.bitcoin.secp256k1_main import *
 import hmac
 import hashlib
 from binascii import hexlify
