--- rescan-script.bat
@@ -1,3 +1,0 @@
-@echo off
-python rescan-script.py
-pause
